requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pyyaml==6.0.2
requests==2.32.3
streamlit==1.37.1
